Add clang-format and clang-tidy configuration files

+++ src/rqt_multiplot/CurveConfigWidget.cpp
@@ -39,12 +39,9 @@ CurveConfigWidget::CurveConfigWidget(QWidget* parent)
       messageTopicRegistry_(new MessageTopicRegistry(this)) {
   ui_->setupUi(this);
 
-  ui_->pushButtonCopyRight->setIcon(
-      QIcon(packageResourcePath("resource/22x22/arrow_right.png")));
-  ui_->pushButtonCopyLeft->setIcon(
-      QIcon(packageResourcePath("resource/22x22/arrow_left.png")));
-  ui_->pushButtonSwap->setIcon(
-      QIcon(packageResourcePath("resource/22x22/arrows_right_left.png")));
+  ui_->pushButtonCopyRight->setIcon(QIcon(packageResourcePath("resource/22x22/arrow_right.png")));
+  ui_->pushButtonCopyLeft->setIcon(QIcon(packageResourcePath("resource/22x22/arrow_left.png")));
+  ui_->pushButtonSwap->setIcon(QIcon(packageResourcePath("resource/22x22/arrows_right_left.png")));
 
   ui_->curveAxisConfigWidgetX->setConfig(config_->getAxisConfig(CurveConfig::X));
   ui_->curveAxisConfigWidgetY->setConfig(config_->getAxisConfig(CurveConfig::Y));
@@ -130,7 +127,7 @@ void CurveConfigWidget::configTitleChanged(const QString& title) {
 }
 
 void CurveConfigWidget::configSubscriberQueueSizeChanged(size_t queueSize) {
-  ui_->spinBoxSubscriberQueueSize->setValue(queueSize);
+  ui_->spinBoxSubscriberQueueSize->setValue(static_cast<int>(queueSize));
 }
 
 void CurveConfigWidget::configAxisConfigTopicChanged(const QString& topic) {
