+++ src/rqt_multiplot/PlotTableConfigWidget.cpp
@@ -47,8 +47,7 @@ PlotTableConfigWidget::PlotTableConfigWidget(QWidget* parent)
   ui_->pushButtonRun->setIcon(QIcon(packageResourcePath("resource/16x16/run.png")));
   ui_->pushButtonPause->setIcon(QIcon(packageResourcePath("resource/16x16/pause.png")));
   ui_->pushButtonClear->setIcon(QIcon(packageResourcePath("resource/16x16/clear.png")));
-  ui_->pushButtonImportExport->setIcon(
-      QIcon(packageResourcePath("resource/16x16/eject.png")));
+  ui_->pushButtonImportExport->setIcon(QIcon(packageResourcePath("resource/16x16/eject.png")));
 
   ui_->pushButtonPause->setEnabled(false);
 
@@ -194,8 +193,8 @@ void PlotTableConfigWidget::configForegroundColorChanged(const QColor& color) {
 }
 
 void PlotTableConfigWidget::configNumPlotsChanged(size_t numRows, size_t numColumns) {
-  ui_->spinBoxRows->setValue(numRows);
-  ui_->spinBoxColumns->setValue(numColumns);
+  ui_->spinBoxRows->setValue(static_cast<int>(numRows));
+  ui_->spinBoxColumns->setValue(static_cast<int>(numColumns));
 }
 
 void PlotTableConfigWidget::configLinkScaleChanged(bool link) {
